requirements.txt
numpy~=1.23.3
opencv-python~=4.6.0.66
Pillow~=9.2.0
PyAutoGUI~=0.9.53
pytesseract~=0.3.10
Shapely~=1.8.4